pyproject.toml
[project]
name = "knowledge-mcp"
version = "0.0.1"
description = "Local MCP server providing hybrid (BM25 + vector) retrieval over a markdown knowledge vault."
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "Issei Kuzuki" }]
keywords = ["mcp", "rag", "knowledge-base", "markdown", "claude"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-frontmatter>=1.1",
    "watchdog>=4.0",
    "blake3>=0.4",
    "langchain-text-splitters>=0.2",
    "tiktoken>=0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[project.scripts]
knowledge-mcp = "knowledge_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["knowledge_mcp"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
